Получение звонков вынесено в репозиторий
