+++ VodovozBusiness/EntityRepositories/Roboats/RoboatsRepository.cs
@@ -472,5 +472,24 @@ namespace Vodovoz.EntityRepositories.Roboats
 				return result;
 			}
 		}
+
+		public IEnumerable<RoboatsCall> GetStaleCalls(IUnitOfWork uow)
+		{
+			RoboatsCall roboatsCallAlias = null;
+			var staleCalls = uow.Session.QueryOver(() => roboatsCallAlias)
+				.Where(() => roboatsCallAlias.CallTime < DateTime.Now.AddMinutes(-_roboatsSettings.CallTimeout))
+				.Where(() => roboatsCallAlias.Status == RoboatsCallStatus.InProgress)
+				.List();
+			return staleCalls;
+		}
+
+		public RoboatsCall GetCall(IUnitOfWork uow, Guid callGuid)
+		{
+			RoboatsCall roboatsCallAlias = null;
+			var call = uow.Session.QueryOver(() => roboatsCallAlias)
+				.Where(() => roboatsCallAlias.CallGuid == callGuid)
+				.SingleOrDefault();
+			return call;
+		}
 	}
 }
